rider-app/refactor: decompose Engine.hs so a second WhatsApp flow can be built beside booking
